requirements.txt
paho-mqtt
numpy
psutil